move string algorithms to new directory
